types(lib): export types with lib
