+++ src/components/Button/types.ts
@@ -1,0 +1,5 @@
+export interface PdapButtonProps {
+	className?: string;
+	intent?: 'primary' | 'secondary';
+	isLoading?: boolean;
+}
