+++ src/components/FlexContainer/types.ts
@@ -1,0 +1,5 @@
+export interface PdapFlexContainerProps {
+	alignment?: 'center' | 'start';
+	className?: string;
+	component?: string;
+}
