+++ src/components/Footer/types.ts
@@ -1,0 +1,9 @@
+export interface PdapFooterSocialLinks {
+	to: string;
+	text: string;
+}
+
+export interface PdapFooterProps {
+	logoImageSrc?: string;
+	logoImageAnchorPath?: string;
+}
