+++ src/components/GridContainer/GridContainer.vue
@@ -7,14 +7,7 @@
 <script setup lang="ts">
 import { reactive } from 'vue';
 
-export interface PdapGridContainerProps {
-	className?: string;
-	columns?: 1 | 2 | 3 | 'auto';
-	component?: string;
-	rows?: number | 'auto';
-	templateColumns?: string;
-	templateRows?: string;
-}
+import { PdapGridContainerProps } from './types';
 
 const props = withDefaults(defineProps<PdapGridContainerProps>(), {
 	columns: 'auto',
