+++ src/components/GridItem/types.ts
@@ -1,0 +1,6 @@
+export interface PdapGridItemProps {
+	className?: string;
+	component?: string;
+	spanColumn?: 1 | 2 | 3;
+	spanRow?: number;
+}
