+++ src/components/Header/types.ts
@@ -1,0 +1,5 @@
+export interface PdapHeaderProps {
+	className?: string;
+	logoImageSrc?: string;
+	logoImageAnchorPath?: string;
+}
