+++ src/components/TileIcon/types.ts
@@ -1,0 +1,4 @@
+export interface PdapTileIconProps {
+	imgAltText: string;
+	imgSrc: string;
+}
