+++ src/index.ts
@@ -3,3 +3,15 @@ export * from './components';
 
 // Styles
 import './styles/styles.css';
+
+// Types
+export * from './components/Button/types';
+export * from './components/FlexContainer/types';
+export * from './components/Footer/types';
+export * from './components/Form/types';
+export * from './components/GridContainer/types';
+export * from './components/GridItem/types';
+export * from './components/Header/types';
+export * from './components/Input/types';
+export * from './components/Nav/types';
+export * from './components/TileIcon/types';
